{{cookiecutter.project_slug}}/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-filter>=25.2",
    "django-cors-headers>=4.9.0",
    "django-allauth[headless,idp-oidc]>=65.15.1",
    "django-unfold>=0.89.0",
    "django-storages>=1.14.6",
    "boto3>=1.42.87",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "psycopg[binary]>=3.3.3",
    "whitenoise>=6.12.0",
    "dotenv>=0.9.9",
]

[tool.ruff]
fix = true
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
exclude = ["/*/migrations/**"]
